fix: before initialization 에러 수정
