+++ apps/web/src/components/market/marketFilter/marketFilter.tsx
@@ -10,7 +10,7 @@ import theme from 'constants/theme';
 import { useRouter } from 'next/router';
 import * as React from 'react';
 import { css } from 'styled-components';
-import { convertQuery, makeFilterLabel } from 'utils/market/marketFilter';
+import { convertQuery, makeFilterLabel } from 'utils/market/marketQuery';
 
 import Close from '../../../assets/svg/close.svg';
 import Refresh from '../../../assets/svg/refresh.svg';
