+++ apps/web/src/components/market/marketSelect/marketSelect.tsx
@@ -1,7 +1,7 @@
 import Select from 'components/common/select';
 import { useRouter } from 'next/router';
 import { SelectType } from '@supercarmarket/types/market';
-import { makeSelectQuery } from 'utils/market/marketFilter';
+import { makeSelectQuery } from 'utils/market/marketQuery';
 
 import * as Styled from './marketSelect.styled';
